Scheme lambda type.
A lambda procedure is an anonymous procedure consisting of a set of
experssions to be evaluated sequentially every time the procedure is
invoked.

It can accept arguments. When initializing a lambda procedure, you can
specify argument identifiers that will be substituted by the provided
arguments when running the procedure. You can also specify optional
arguments with default values, and a rest argument that will be replaced
with a list containing every argument not assigned to a previous
identifier.